feat: add better response error typing
