+++ src/services/RegisterService.ts
@@ -1,10 +1,10 @@
-import { langApi, User } from "../generated/services/langApi";
-import IRegisterService from "./interfaces/IRegisterService";
-import IRegisterCreateParams from "./interfaces/IRegisterCreateParams";
-import store from "../model/store";
 import { injectable } from "inversify";
+import { langApi, User } from "../generated/services/langApi";
 import { setRegisterMessage, successRegister } from "../model/feature/auth";
-import { isKnownError, isSerializedError, isSuccessResponse } from "./utils/ErrorResposne";
+import store from "../model/store";
+import IRegisterCreateParams from "./interfaces/IRegisterCreateParams";
+import IRegisterService from "./interfaces/IRegisterService";
+import { isCustomError, isFetchError, isParsingError, isSerializedError, isSuccessResponse, isUnknownError } from "./utils/ErrorResposne";
 
 @injectable()
 export default class RegisterService implements IRegisterService {
@@ -17,17 +17,24 @@ export default class RegisterService implements IRegisterService {
         return true;
       }
 
+      const { error  } = response;
+
       let message: string = '';
 
-      if (isSerializedError(response.error)) {
-        message = response.error.message ?? 'Неизвестная ошибка';
-      } else if (isKnownError(response.error)) {
-        message = response.error.error;
-      } else {
+      if (isSerializedError(error)) {
+        message = error.message ?? 'Неизвестная ошибка';
+      } else if (isUnknownError(error)) {
         message = 'Неизвестная ошибка';
+      } else if (isParsingError(error)) {
+        message = error.data ?? 'Неизвестная ошибка'
+      } else if (isFetchError(error) || isCustomError(error)) {
+        message = error.error;
+      } else {
+        throw Error('Undefined behavior');
       }
 
       store.dispatch(setRegisterMessage(message));
+
       return false;
     })
   }
