+++ src/view/pages/Main/index.tsx
@@ -5,7 +5,7 @@ import Box from '@mui/material/Box';
 import AppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import Button from '@mui/material/Button';
-import AuthModal from '../../../components/AuthModal';
+import AuthModal from '../../components/AuthModal';
 import { useSelector } from "react-redux";
 import { RootState } from "../../../model/store";
 import userSlice from "../../../model/feature/user";
